chore(git-properties): renames git properties to make it more intuitive.
